Refactor Request Telemetry Cache
